some quick refactors
